+++ src/podofo/main/PdfAnnotationActionBase.h
@@ -102,7 +102,7 @@ namespace PoDoFo {
          *
          *  \see GetAction
          */
-        void SetAction(const std::shared_ptr<PdfAction>& action);
+        void SetAction(const nullable<const PdfAction&>& action);
 
         /** Get the action that is executed for this annotation
          *  \returns an action object. The action object is owned
@@ -110,13 +110,18 @@ namespace PoDoFo {
          *
          *  \see SetAction
          */
-        std::shared_ptr<PdfAction> GetAction() const;
+        nullable<PdfAction&> GetAction();
+        nullable<const PdfAction&> GetAction() const;
+
+    protected:
+        virtual void OnActionSet();
+        void ResetAction();
 
     private:
-        std::shared_ptr<PdfAction> getAction();
+        nullable<PdfAction&> getAction();
 
     private:
-        std::shared_ptr<PdfAction> m_Action;
+        nullable<std::unique_ptr<PdfAction>> m_Action;
     };
 }