+++ src/podofo/main/PdfAnnotation_Types.cpp
@@ -13,54 +13,94 @@
 using namespace std;
 using namespace PoDoFo;
 
-void PdfAnnotationLink::SetDestination(const shared_ptr<PdfDestination>& destination)
+void PdfAnnotationLink::SetDestination(const nullable<const PdfDestination&>& destination)
 {
-    destination->AddToDictionary(GetDictionary());
-    m_Destination = destination;
+    if (destination == nullptr)
+    {
+        GetDictionary().RemoveKey("Dest");
+        m_Destination = { };
+    }
+    else
+    {
+        m_Destination = unique_ptr<PdfDestination>(new PdfDestination(*destination));
+        ResetAction();
+        destination->AddToDictionary(GetDictionary());
+    }
 }
 
-shared_ptr<PdfDestination> PdfAnnotationLink::GetDestination() const
+nullable<PdfDestination&> PdfAnnotationLink::GetDestination()
+{
+    return getDestination();
+}
+
+nullable<const PdfDestination&> PdfAnnotationLink::GetDestination() const
 {
     return const_cast<PdfAnnotationLink&>(*this).getDestination();
 }
 
-shared_ptr<PdfDestination> PdfAnnotationLink::getDestination()
+nullable<PdfDestination&> PdfAnnotationLink::getDestination()
 {
     if (m_Destination == nullptr)
     {
         auto obj = GetDictionary().FindKey("Dest");
         if (obj == nullptr)
-            return nullptr;
-
-        m_Destination.reset(new PdfDestination(*obj));
+            m_Destination = { };
+        else
+            m_Destination = unique_ptr<PdfDestination>(new PdfDestination(*obj));
     }
 
-    return m_Destination;
+    if (*m_Destination == nullptr)
+        return { };
+    else
+        return **m_Destination;
 }
 
-void PdfAnnotationFileAttachement::SetFileAttachement(const shared_ptr<PdfFileSpec>& fileSpec)
+void PdfAnnotationLink::OnActionSet()
 {
-    GetDictionary().AddKey("FS", fileSpec->GetObject().GetIndirectReference());
-    m_FileSpec = fileSpec;
+    m_Destination = { };
+    GetDictionary().RemoveKey("Dest");
 }
 
-shared_ptr<PdfFileSpec> PdfAnnotationFileAttachement::GetFileAttachement() const
+void PdfAnnotationFileAttachment::SetFileAttachment(const nullable<PdfFileSpec&>& fileSpec)
 {
-    return const_cast<PdfAnnotationFileAttachement&>(*this).getFileAttachment();
+    if (fileSpec == nullptr)
+    {
+        m_FileSpec = { };
+        GetDictionary().RemoveKey("FS");
+    }
+    else
+    {
+        GetDictionary().AddKeyIndirect("FS", fileSpec->GetObject());
+        m_FileSpec = unique_ptr<PdfFileSpec>(new PdfFileSpec(*fileSpec));
+    }
 }
 
-shared_ptr<PdfFileSpec> PdfAnnotationFileAttachement::getFileAttachment()
+nullable<PdfFileSpec&> PdfAnnotationFileAttachment::GetFileAttachment()
+{
+    return getFileAttachment();
+}
+
+nullable<const PdfFileSpec&> PdfAnnotationFileAttachment::GetFileAttachment() const
+{
+    return const_cast<PdfAnnotationFileAttachment&>(*this).getFileAttachment();
+}
+
+
+nullable<PdfFileSpec&> PdfAnnotationFileAttachment::getFileAttachment()
 {
     if (m_FileSpec == nullptr)
     {
         auto obj = GetDictionary().FindKey("FS");
         if (obj == nullptr)
-            return nullptr;
-
-        m_FileSpec.reset(new PdfFileSpec(*obj));
+            m_FileSpec = { };
+        else
+            m_FileSpec = unique_ptr<PdfFileSpec>(new PdfFileSpec(*obj));
     }
 
-    return m_FileSpec;
+    if (*m_FileSpec == nullptr)
+        return { };
+    else
+        return **m_FileSpec;
 }
 
 void PdfAnnotationPopup::SetOpen(const nullable<bool>& value)
@@ -139,12 +179,12 @@ PdfAnnotationCaret::PdfAnnotationCaret(PdfObject& obj)
 {
 }
 
-PdfAnnotationFileAttachement::PdfAnnotationFileAttachement(PdfPage& page, const Rect& rect)
+PdfAnnotationFileAttachment::PdfAnnotationFileAttachment(PdfPage& page, const Rect& rect)
     : PdfAnnotation(page, PdfAnnotationType::FileAttachement, rect)
 {
 }
 
-PdfAnnotationFileAttachement::PdfAnnotationFileAttachement(PdfObject& obj)
+PdfAnnotationFileAttachment::PdfAnnotationFileAttachment(PdfObject& obj)
     : PdfAnnotation(obj, PdfAnnotationType::FileAttachement)
 {
 }