+++ src/podofo/main/PdfDestination.h
@@ -50,7 +50,10 @@ enum class PdfDestinationType
  */
 class PODOFO_API PdfDestination final : public PdfArrayElement
 {
-public:
+    friend class PdfDocument;
+    friend class PdfAnnotationLink;
+    friend class PdfOutlineItem;
+
     /** Create a new PdfDestination from an existing PdfObject (such as loaded from a doc)
      *  \param obj the object to construct from
      */
@@ -60,17 +63,22 @@ public:
      */
     PdfDestination(PdfDocument& doc);
 
+    PdfDestination(const PdfDestination&) = default;
+
+public:
+    static bool TryCreateFromObject(PdfObject& obj, std::unique_ptr<PdfDestination>& dest);
+
     /** Create a new PdfDestination with a page as destination
      *  \param page a page which is the destination
      *  \param fit fit mode for the page. Must be PdfDestinationFit::Fit or PdfDestinationFit::FitB
      */
-    PdfDestination(const PdfPage& page, PdfDestinationFit fit = PdfDestinationFit::Fit);
+    void SetDestination(const PdfPage& page, PdfDestinationFit fit = PdfDestinationFit::Fit);
 
     /** Create a destination to a page with its contents magnified to fit into the given rectangle
      *  \param page a page which is the destination
      *  \param rect magnify the page so that the contents of the rectangle are visible
      */
-    PdfDestination(const PdfPage& page, const Rect& rect);
+    void SetDestination(const PdfPage& page, const Rect& rect);
 
     /** Create a new destination to a page with specified left
      *  and top coordinates and a zoom factor.
@@ -79,7 +87,7 @@ public:
      *  \param top top coordinate
      *  \param zoom zoom factor in the viewer
      */
-    PdfDestination(const PdfPage& page, double left, double top, double zoom);
+    void SetDestination(const PdfPage& page, double left, double top, double zoom);
 
     /** Create a new destination to a page.
      *  \param page a page which is the destination
@@ -87,11 +95,8 @@ public:
      *              PdfDestinationFit::FitV, PdfDestinationFit::FitBH, PdfDestinationFit::FitBV
      *  \param value value which is a required argument for the selected fit mode
      */
-    PdfDestination(const PdfPage& page, PdfDestinationFit fit, double value);
+    void SetDestination(const PdfPage& page, PdfDestinationFit fit, double value);
 
-    static std::unique_ptr<PdfDestination> Create(PdfObject& obj);
-
-public:
     /** Get the page that this destination points to
      *  Requires that this PdfDestination was somehow
      *  created by or from a PdfDocument. Won't work otherwise.
@@ -146,6 +151,7 @@ public:
      */
     double GetDValue() const;
 
+private:
     /** Adds this destination to an dictionary.
      *  This method handles the all the complexities of making sure it's added correctly
      *