+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Fixed PdfStreamedDocument, see #88
 - Added high-level signing API, see PdfSignerCMS and "TestSignature1" test case
 - Set PdfSignature to have correct /ByteRange and /Contents after signing with PoDoFo::SignDocument
+- Reviewed PdfFileSpec, PdfAction, PdfDestination API and their usage in
+PdfOutlineItem, PdfOutlines, PdfAnnotationActionBase, PdfAnnotationLink PdfAnnotationFileAttachment
 
 ## Version 0.10.3
 - Fixed big performance regression introduced in 0.10, see #108